requirements.txt
streamlit>=1.33.0
openai>=1.30.1
python-dotenv>=1.0.0